add AwaitilityExtensionModule
Both AwaitilitySupport (meant to be used as @Mixin) and
AwaitilityTrait (meant to be implemented) have one big flaw. When not used
for one test class, this test cases might or might not fail, depending on
execution order.

The AwaitilityExtensionModule will automatically add a new method
until(Closure) to the ConditionFactory. So with groovy 2 and above, there
is effectively no need any more to use AwaitilitySupport, AwaitilityTrait
or AwaitilityGroovyBridge. All you need to do is add awaitility-groovy
to your classpath and you are good.